Fixed Fog::AWS::SimpleDB#delete_attributes.
Deleting attributes was not working in mock mode. There were two
bugs in this function.The first was that attributes were being
deleted from the domain as opposed to the item. The second was
that when no attributes were passed, which should delete all
attributes, no attributes were actually being deleted. I've written
tests that recreate these bugs and I have included a fix as well.