* Changed LoadError to Fog::Error::LoadError when missing configuration * When running from command line, rescue the exception, and print the help message
When running fog for the first time with no setup, the missing_exception message
was surrounded above and below with a stack trace, which is not particularly
user-friendly. Since we've rescuing the LoadError only in bin/fog, we know
we don't have to worry about higher level libraries not getting the exception,
since it wouldn't propagate between processes.

Changing LoadError to Fog::Error::LoadError lets us know that this was
an issue on fog's part, rather than some other library we might be
using. Since we inherit from the LoadError class, any higher-level
libraries that were using a rescue based on LoadError will still work.