Added Fog::CurrentMachine#ip_address.
This function allows you to get the ip address of the current machine. I
found this useful because I wanted to add my personal or production
machine to a specific RDS security group. The call is quite simple and
connects to amazon's checkip website. The class is threadsafe and I have
included specs with it as well. I have also added the
Fog::AWS::RDS::SecurityGroup#authorize_me function to make adding the
current machine to a given RDS security group very easy.