(#9241) Add skeleton VMware vSphere platform support
This patch adds a compute service to fog setting the stage to model
VMware virtual machines using Fog.  The patch adds support for:

    rdebug -- fog vsphere
    >>> connection = Fog::Compute.new(:provider => :vsphere)

The connection to the VMware API is implemented along with
authentication using an username and password.  The connection is not
fully secured with this patch because no validation of the SSL
certificate is implemented.

Raw API requests are working with this patch, but none of the API
requests have associated Fog models or collections.