Tweak to escape the Cloud Files filename before passing to public_url.
Previously, a Rackspace Cloud Files storage object with a name like
"my_files/1234/original/picture #1.jpg" would generate a public URL like:

http://c21641.r41.cf1.rackcdn.com/user_files/1313284/thumb/IMG_4034 #1.jpg

The # character in the URL would get interpreted by Cloud Files as an
anchor marker, and the file would not load.  This change does the
Fog::Rackspace.escape on the key (excluding the / path separator character)
to generate a working URL of the form:

http://c21641.r41.cf1.rackcdn.com/user_files/1313284/thumb/IMG_4034%20%231.jpg

Which is correctly interpreted by Cloud Files and pulls up.

Both of those URLs are live to demonstrate the issue.