[vcloud|compute] improve models add tests
Improve the models to work with the current api version. add tests
that use fake-data, that was captured from a real vCloud.

Also removed a couple of dead code. Tests have been added for
everything that changed or have been added. All the existing but
untouched parts have not been tested.