[core] `@credential` should always be a symbol
Similar to #504, say your ~/.fog had multiple stanzas:

```yaml
:default:
  :aws_access_key_id: aaa
test:
  :aws_access_key_id: bbb
:prod:
  :aws_access_key_id: ccc
```

By default, fog will look for either the credential *symbol* `:default` or for the *string* stored in `FOG_CREDENTIAL` (in `lib/fog/core/credentials.rb` method `self.credential`). When the fog credentials file given above is read in, it’s passed through `self.symbolize_credentials` which makes *symbols* out of all the keys (so `prod`, `test`, and `default` will all be symbols) and saves the list to credentials then looks for `credentials[credential]` which is indexing a string in a hash of symbols.

Any value that `@credential` is set to should be a symbol, because the hash keys will only be symbols.