Fix linked clone mocked test unhandled exception
Without this patch, the test for the vSphere linked clone option to the
vm_clone request throws an exception.  For the full exception please see
GH-697 discussion comments on Github at [1]

This patch fixes the issue by removing the cleanup code which is not
necessary for a Mocked test.  The cleanup code was calling a method on
the response_linked object which is not actually in scope.

[1] https://github.com/fog/fog/pull/697