Add source for getting instance mac address
Previous to this commit, the only source to obtain the mac address for
the primary interface on an instance was through libvirt directly.  If
libvirt was unable to retrieve the mac address, it would be nil.  This
commit adds the ability to dump the XML using the virsh command and
attempting to retrieve the mac address from that.