Do not hardcode the avl_zone order for hp compute.
HP compute avl zones were hardcoded to assume that the first response
is for az-1, then az-2 (and not support for az-3). Change this to look
up the avl zone based on the passed in name. The passed in avl zones
should now be strings of the form "az-1", "az-2", etc, as opposed to
symbols that are currently used.