Submit password/ssh_key/username through POST body
Password or ssh_key, and username are currently submitted through the
request's querystring, which means they will appear in server logs.

This change moves the sensitive parameters to the request body, adds
client-side checking of the required parameters, and corrects the method
documentation to reflect addition of location_id parameter and specify
the parameter types as strings (they are alpha-numeric UUIDs).