[aws|ec2] describe_availability_zones parser handles nested <item> tags
This patch fixes the parser so there is no empty {'messageSet' => []}
value.

At the moment, AWS is returning a body like that is incorrectly parsed:

     <availabilityZoneInfo>
        ...
        <item>
            <zoneName>us-east-1b</zoneName>
            <zoneState>impaired</zoneState>
            <regionName>us-east-1</regionName>
            <messageSet>
                <item>
                    <message>We have restored the ability to launch new
EC2 instances in this Availability Zone. At this point, customers
should be able to launch instances in any Availability Zone in the
US-EAST-1 region. We are continuing to restore impaired volumes and
their attached instances.</message>
                </item>
            </messageSet>