[Brightbox] Deprecates overloaded requests
Fog::Compute::Brightbox::Real#get_user and #get_account were overloaded
to return the currently authenticated user or the scoped account
respectively if no identifier was passed.

This is strictly wrong because the requests should map to one API
endpoint but these now switch between two.

Behaviour is deprecated and will be removed at a suitable point in the
future.