[docs] Replaces link to API to `rubydoc.info`
For historical reasons we generated and hosted our own versions of the
API docs, mostly so we could keep multiple versions.

http://rubydoc.info is giving this to the community for free amd also
works with github to offer the latest docs, so we have decided to stop
generating and hosting our own API docs.

This changes the link on `fog.io` to reference `rubydoc.info` instead
and removes the uploading rake task.

It does not remove the `yard` rake task so documentation can still be
generated locally.

Fixes #1274