[Brightbox] Adds means to update scoped account
A Compute instance encapsulates a connection for a client to the
Brightbox API. Users can have multiple accounts but there was no easy
way to switch between them when account had to be passed in via the
initializer.

Now the scoped account can be set on an existing instance which overrides
any configured setting but it can be reset if needed.

The #request method still can accept `account_id` as an option which
again overrides the previous settings.

Finally the parameter is now correctly sent as a query string parameter
not part of the API request JSON.