OpenStack create_server mocks now match reality
Previously the mocks would return the same data as list_servers_detail.
OpenStack does not return all of this data so you must reload a server
after creation to get the necessary data when using fog for real.

Now the mock returns the same detail as the real call with the extra
detail stored in the mock data for retrieval by list_servers_detail.