[core] Adds #persisted? to Fog models
Fog models have been able to use #new_record? for years to use an
ActiveRecord like interface to check for peristence of a resource.

When Rails 3 switched to using ActiveModel API this changed to
be #persisted?

This deprecates #new_record? and adds #persisted? as the first little
step to making fog models easier to use using ActiveModel API.

As discussed on the tickes, ActiveModel is not planned to become a
direct dependency for fog.

Relates to #1276