[ecloud] Fixes tests by duplicating test setup
In order to have two different groups of server tests the ecloud tests
are in two sets within the one file - which explains the setup outside
of the Shindo block.

Shindo's tags do not work when nested so this fixes the problem by
duplicating the test setup in the second group of tests.

Fog::Compute can't safely sit outside the top Shindo block so the
duplication is preferable to breaking all/ecloud tests.

Also corrected the missing provider tag in this second set so the
"attributes" tests do not run when should be filtered by tags.