Disables tests with race conditions affecting CI
GH#1546 - The new build system on Travis CI is erring amongst three
tests. Initial investigating hints that they contain race conditions
in the tests (rather than the code).

We've decided to disable the tests whilst we look to resolve since it is
devaluing the CI for the rest of the project.