Fix circular requires in fog
When run with -w fog prints out 1763 lines of warnings, most of which
are:

    warning: loading in progress, circular require considered harmful - [file]

followed by a backtrace showing the circular require path.

Each of these circular requires are pulling in the fog provider
namespace, but due to the explicit namespacing in fog this is not
necessary.

The require may be necessary to have the service properly registered
with the provider.  Without this require this patch may break a user who
requires "fog/openstack/compute", but I am unsure if this is a supported
use of fog, please advise.