Remove spurious warning...
This removes a warning that is outputed by Fog when a S3 bucket name
contains a period. This resulted in excessive logging of a condition when
using Fog to push to a S3 bucket configured to serve a web site. See
fredjean/middleman-s3_sync#10 to see the issue that triggered this commit.