Ease the restriction on version of nokogiri
We want to be able to use Nokogiri 1.6.0 in a project that also requires
Fog. This is because of a problem with earlier versions of Nokogiri and
LibXML 2.9[1]. We ran all the Fog tests against Nokogiri 1.6.0 and
everything passes.

The Fog README states that it should work with Ruby 1.8.7 and Ruby 1.9
but  Nokogiri 1.6.0 will complain if you try to use it with anything
less than Ruby 1.9.2. This might cause problems if we want to get this
change pushed back to Fog.

[1] https://github.com/sparklemotion/nokogiri/issues/829