Add RDS API version parameter option.
RDS API has moved to 2013-05-15 and added additional methods and
parameters to existing ones (such as NewDBInstanceIdentifier in
ModifyDBInstance). It makes sense to use the new API version, however
it requires more testing and possibly implementing additional methods.
This commit allows the version to be passed when instantiating a fog
object, so that this can be done piece-meal, without blocking existing
fog users.

Refs: #1931