Infer the 'image' URL correctly when inserting a server.
When inserting (creating) a new server, using a private (project
specific) image, the code would infer the resource URI for the
image incorrectly, because it would try to find the image in the
current project using an incorrect method to access the name of
the project. This patch fixes the variable reference, and changes
the loop logic to try all known project names in the same fashion.

Refs #1933