Query global projects when get/list'ing compute images.
This patch refactors some of the code that would query global
projects when get'ing images. It makes the list of global projects
a constant on the Images collection, so that both 'get' and 'list'
can use the same list of projects to query.

Also, when bootstrapping/create'ing a server, the validation of
the specified image name is done by trying to 'get' the image
instead of 'save'ing it.