[AWS|ASG] filter mocked results for describe_auto_scaling_policies
This is the same issue as was fixed for ASGs in #2040 / de274452

Because Policies#get uses this method to fetch a single Policy, it was
always returning the first Policy in the data set, not the one that
actually matched the specified Policy name. Likewise, providing a filter
for `AutoScalingGroupName` did not actually filter the returned data
from this call when mocked.