[openstack|compute] Basic examples for Compute
This contains some basics examples on how to set up a VM, do power
operations, and find floating IP pools. It should be enough for a
beginner to get started and familiar with Fog so that more complex
things can be searched by the user.

Get list of image names