[Brightbox] Fix test issue with reusing servers
To save time, we hoped to use an existing server to read the interface
JSON and it did a naive selection of the first server.

Server's that are deleted have two relevant characteristics. 1) They no
longer have interfaces and 2) they are reported in the output in their
deleted state for an hour after being removed.

Since the test was not filtering for active servers, subsequent runs
could pick up a deleted server and fail to find the identifier for it's
missing first interface.

This simplifies matters by creating a new server for the test.