[vcloud_director] Handle Guest admin password
Admin passwords can be retrieved and modified using the VMWare console,
so it's useful to allow the same operations here.

Moreover, when an admin password is enabled and "admin password auto" is
false, it's impossible to modify Guest settings.
i.e., "The administrator password cannot be empty when it is enabled and
automatic password generation is not selected."

This patch also fixes `put_guest_customization_section_vapp` where an
AdminPassword element was never created.