[vcloud_director] get_network_complete
The existing get_network request method does not return the complete
network details provided by the API -- the parser is lossy. In addition, it
rewrites vCloud elements in a ruby-ish way, which then requires retranslation
to load back in the API.

It is not possible to adjust the get_network request without breaking code that
already depends on it (in particular, the current Network model), hence the need
for a tandem method.

The get_network_complete method returns in a format that should be loadable
into post_create_org_vdc_network, as suggested by API documentation.