Make Coveralls opt-in
Due to looking for a value of 'false' whenever I run tests locally
without the `COVERAGE` env variable coverage is very slowly, still
calculated - it then aborts without doing anything because I'm not using
a CI setup.

This adds the setting in for Travis to opt in and should exclude
coverage work for anyone running tests locally.