Allow for specification of vcloud_token via ENV
This change permits the use of an external login method for vCloud
Director, relying on the fact that the password is only needed to
create a session in the API, subsequently accessed via the token.

In effect, this means that users do not need to store a valid
password in the FOG_RC, significantly reducing the risk footprint.

The default token lifetime is '30 minutes idle' - any activity
extends the life by another 30 mins. This makes this a very workable
solution for general user sessions.