support rackspace storage delete_at and delete_after headers
Rackspace Cloud Files allows the caller to set a time in the future
after which the uploaded asset will be automatically deleted.

http://docs.rackspace.com/files/api/v1/cf-devguide/content/Storage_Object_Services-d1e4300.html

This commit adds `deleted_at` and `deleted_after` attributes to the
`rackspace/models/storage/file.rb`. They are then added to the HTTP
headers when requested.

fix typo

add marshall yount to the contributors file