Fix Joyent service declarations
When the Joyent provider was merged in it was after a reworking of how
requires and services were declared.

This meant the call to `service` was no longer correct.

This updates the signature and the mocked tests now pass (locally at
least) however the structuring and placement of the files may not be
inconsistent with the results of [GH-1712]