[Brightbox] Destroy snapshot after completion
This works around an issue where you can not detect if a Cloud SQL
Instance is being snapshotted.

A new API restriction is that you can no longer delete an instance that
is being snapshotted so this suddenly began failing.

Now we look for a new snapshot and poll until it is ready only then can
we safely issue the destroy command.