Add network model tests
Largely based on the vapp tests, except for explicitly
trying to find a suitable network to test against rather than
just blindly picking the first.

Test lazy loading is working, and that values loaded match
our data from the raw service request.

Lazy loading unfortunately complicates testing as per the
current implementation - determining that the loading has
happened is on a field-by-field basis, which leaves fields
that are undefined in vCloud in a 'NonLoaded' state. Possibly worth
revisiting this logic in a separate commit.