[core] Deprecate Fog::Connection
This reapplies the deprecation warnings that were originally part of
7ee3535 when using Fog::Connection request.

Unfortunately the Mock world does not exercise any of the code affected
by this change so I missed that I had adjusted the signature of the
method.

If you pass a parser argument into Fog::Connection you should switch to
using Fog::XML::SAXParserConnection and pass the parser as the first
argument to `#request`

If you don't it is better to use Fog::Core::Connection to get rid of the
backwards compatibility code and XML dependency.

This will make fog VERY NOISY since you'll get output for every request
on out of date services.