[DigitalOcean] Skip consistently timing out tests
The server tests in digital ocean have been frequently timing out.

* https://travis-ci.org/fog/fog/jobs/20605301#L1030
* https://travis-ci.org/fog/fog/jobs/19765753#L1233
...and many more

Each timeout run wastes Travis resources and the time of the people who
have to check why their unrelated changes are failed.

We've even become so complacent about it, standard practice is to just
comment and merge.

So I'm disabling the tests until someone fixes it or we replace the
server tests with a cleaner approach, whichever comes first.