Removed unicode NFC normalization of S3 object keys.
S3 does not require normalization of S3 object keys and uses strict byte
comparison of object keys, not equivalent unicode character comparisons,
to store and retrieve objects. This means that storing and retrieving objects
with fog would cause the objects to be inaccessible by other libraries,
languages, and systems that don't normalize object keys. Given that there is
no benefit to normalization, except perhaps reducing byte count of object
keys, it ought to be removed.