[Rackspace] Remove circular requires from Storage
The fog DSL automatically requires files when `models` and `collections`
have been declared.

When Metadata is required, it attempted to require both file, files,
directory and directories, both trying to require the same.

When fog/fog-core#32 was done, the ordering changed so that directory
was required first and referenced metadata which referenced Directory.

Fixes fog/fog#2825