[GH-2873] Ensure fog is using fog-core 1.22+
fog-v1.22.0 went out with the dependencies loosely defined so it would
accept fog-core-v1.21.

fog/fog-core@e164cc7 removed a Vsphere helper method than had been
incorrectly defined on the Fog module. This was then extracted into
fog-core.

The fix was applied to `fog` already but the dependencies allow the
incorrect version of fog to run with the cleaned up version of fog-core
whilst they should not.