[vcloud_director] Add :put_network request
The put_network request updates OrgVdcNetwork details:

- name,
- dns properties,
- gateway

There appear to be some limitations with what the API will allow
to be changes (such as IpRanges) but the input/output of the method
is mapped as usual.

This also adds Mock support for the method.

Network tests have been adjusted so that they work in Mock and Real
mode (bar network_metadata, which is to be implemented in Mock).