Adding Tutum support
Tutum provides CAAS, container-as-a-service, for docker. Using this
library you can CRUD docker images, applications/clusters and
containers/servers.

https://docs.tutum.co/reference/api/

This code is modelled after a mixture of digitalocean and fogdocker.

Tutum uses an application to abstract a scaling group of containers.
To create a new container, create an application with num containers = 1

Container/Application start/stop/redeploy are channeled through a generic
*_action request, since the request/response styling is the same.

At this time there is a known issue with the docs regarding the expected
response code for container creation, where the docs do not match the api.