[vsphere] Use server_id in the Volumes collection.
If a reference to the Vsphere::Server object is passed explicitly, to
the Vsphere::Volumes collection, a loop is introduced when attempting
to print out the Vsphere::Server object. The attributes for the server
will include a reference to the Vsphere::Volumes collection, which in
turn includes a reference to the server object.

As opposed to passing a reference to the Vsphere::Server object to
the Vsphere::Volumes collection, this patch modifies the interface to
expect a Vsphere:Server ID. The ID is used to perform any subsequent
lookups of the server object -- if required.

Fixes: #2938