[vsphere] Remove the relative_path attribute.
The <Vsphere::Server>.relative_path attribute depends on having
initialized the <server>.datacenter field (which is lazy loaded).
When the relative_path is computed from the attributes returned by
the underlying rbvmomi call, it is actually an absolute path to the
instance (that is, it includes the Datacenter). However, the path is
meant to be relative to the Datacenter.

The patch proposes the following solution:
* remove the relative_path attribute
* add a <Vsphere::Server>.relative_path method

This works well with the lazy initialization and restores
functionality to methods that rely on the path actually being relative
to the Datacenter.

Fixes: #3041