Show failure for Openstack Storate service_type
The openstack_service_type for the Openstack Storage Service (Swift) currently
defaults to "object-store".  However, if the Swift service is unvailable, the
authenticate_v2 method attempts to call:

    service_type.join ", "

The problem is that the String "object-store" doesn't have a join method.

The new tests show that "object-store" is an invalid service_type, whereas
["object-store"] is a valid service_type.

https://github.com/fog/fog/issues/3112