Update the Openstack Storage service type
Change the Openstack Storage service (Swift) openstack_service_type default value from
"object-store" to ["object-store"].

The problem with "object-store" as an openstack_service_type is that if the
Swift service is unavailable, the authenticate_v2 method will attempt to call:

    service_type.join ", "

This will fail for anything that isn't an array.

Changing the openstack_service_type to an array solves this problem.
Additionally, it keeps the openstack_service_type consistent with all of the
other Openstack services' openstack_service_type default values.

https://github.com/fog/fog/issues/3112