[vsphere] new default dest_folder in vm_clone
fixes fog/fog#3134

We want to allow someone to not set the dest_folder option because they
might not have a desired folder structure and not care. In case they
follow common examples and use a folder called 'templates/' for templates
we don't want to create the new vm in that folder. It is better to
default to the root folder.

Conflicts:
	lib/fog/vsphere/requests/compute/vm_clone.rb