[AWS] Setup a fog keypair only if not supplied.
When calling bootstrap, the :key_name argument is currently ignored
and not passed to the internal _setup_bootstrap() method. This results
in fog creating a keypair and using it for the instance.

The additional parameter -- new_attributes -- is also never actually
used when calling _setup_bootstrap().

The patch removes the new_attributes parameter and uses the key_name
attribute of the passed server object to ascertain whether a default
fog keypair should be created.