Docker: Support for logs/top operations
In order to test this properly, responses from the mock return values
closer to what the API actually returns. Examples are from Remote API
1.18.

downcasing hash keys cannot be performed on all Docker API operations so
now it will only be done where the top element is a hash. Logs, top, and
many others will not return a hash.