nodes need  an optional :type and :condition
http://docs.rackspace.com/loadbalancers/api/v1.0/clb-devguide/content/Nodes-d1e2173.html
Type of node:

PRIMARY – Nodes defined as PRIMARY are in the normal rotation to receive traffic from the load balancer.

SECONDARY – Nodes defined as SECONDARY are only in the rotation to receive traffic from the load balancer when all the primary nodes fail. This provides a failover feature that automatically routes traffic to the secondary node in the event that the primary node is disabled or in a failing state. Note that active health monitoring must be enabled on the load balancer to enable the failover feature to the secondary node.