[Fog 2.0] Remove 1.8.7 support
This is the official commit to drop Ruby 1.8.7 support from fog.

This is a backwards incompatible change (for hopefully obvious reasons)
so will be part of the next MAJOR fog release `v2.0`

Ruby 1.8.7 is no longer supported by the language developers.

Anyone needing to stay on Ruby 1.8.7 such as users of LTS distros should
remain on the `v1.x` series of fog.