All I want for Christmas are logs clean of warnings about HTTPS access to my bucket...
Using S3 to host a static website requires you to use dots in the bucket name.
As a result, the logs from using a gem like middleman-s3_sync are littered with
warnings. middleman-s3_sync uses path_style by default, so the warning isn't
really needed.

This changerRemoves the warning about using dot in bucket names if it is
already configured to use path style. This shows a deliberate (and probably
knowledgeable) choice.