Tests top level classes interface
Currently most providers used by the `fog` binary give themselves a top
level class. This has become the place a lookup table between services
and the classes which implement them.

This adds compatibility tests for the existing behaviour to allow
controlled extraction to modules.

This top level class behaviour is due to be deprecated and removed in
fog 2.0