Remove the "ovirt_ca_cert_store" option
This option is intended to pass a certificate store containing the
accepted CA certificates, but the underlying "rbovirt" library is about
to remove it because it can only be used with HTTP libraries that
directly support these OpenSSL objects. See here for details:

  https://github.com/abenari/rbovirt/pull/55

As an alternative to this "ovirt_ca_cert_store" the already existing
"ovirt_ca_cert_file" can be used, with the same functinality. Actually
the "ovirt_ca_cert_store" parameter isn't currently used in the fog code

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>